# usm: symbols lookup improvement