<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Draft PRs should be prefixed with `[WIP]` in their title.

-->
### What does this PR do?

Improves performances around 2 heavy computational areas:
1. Determining if a single binary is a goland process (for go-tls hooks)
2. Extracting symbols from a given binary

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

### Motivation

Performance improvement (cpu and memory)

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

Benchmarks:
On main
```
BenchmarkInspectProcessIsNotGo
BenchmarkInspectProcessIsNotGo-16    	   75639	     16092 ns/op	   49306 B/op	       4 allocs/op
BenchmarkInspectProcessIsGo
BenchmarkInspectProcessIsGo-16        	       6	 169743725 ns/op	  137376 B/op	    1787 allocs/op
```

After
```
BenchmarkInspectProcessIsNotGo
BenchmarkInspectProcessIsNotGo-16    	   69145	     17026 ns/op	     146 B/op	       4 allocs/op
BenchmarkInspectProcessIsGo
BenchmarkInspectProcessIsGo-16        	      12	  90951788 ns/op	   24587 B/op	     442 allocs/op
```

Determining if a process is golang:
CPU roughly the same (0.05% diff)
Memory allocated down by ~99%
Allocations the same

Extracting symbols from a golang process:
CPU down by ~45%
Memory allocated down by ~80%
Allocations down by ~75%

Syscalls reductions - Reduced ~66% in read syscall (achieved by reading symbols in batched)

Load test
CPU reduction of 25% (including startup)
![image](https://github.com/DataDog/datadog-agent/assets/17148247/95a311a0-cb41-4aba-a5b1-97fda2d859f6)

Memory is stable
if comparing the startup phase, then we gain 7% reduction
![image](https://github.com/DataDog/datadog-agent/assets/17148247/d489cfe8-90a9-4f93-8e5f-f8c212a8e067)

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.